+++ Changes.md
@@ -4,7 +4,9 @@
 Features
 --------
 
-- LightTool : Added manipulator for disk and point light radii.
+- LightTool :
+  - Added manipulator for disk and point light radii.
+  - Added manipulator for cylinder length.
 
 Fixes
 -----